@@ -13,17 +13,20 @@
     "prepare": "husky install"
   },
   "dependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "zustand": "^3.5.10"
   },
   "devDependencies": {
-    "vite": "^2.6.14",
-    "typescript": "^4.5.4",
+    "@types/react": "^19.2.0",
+    "@types/react-dom": "^19.2.0",
+    "@vitejs/plugin-react": "^1.3.2",
     "eslint": "^7.32.0",
-    "prettier": "^2.5.1",
     "husky": "^7.0.4",
     "lint-staged": "^12.1.2",
+    "prettier": "^2.5.1",
+    "typescript": "^4.5.4",
+    "vite": "^2.6.14",
     "vitest": "^0.0.112"
   },
   "keywords": [
