@@ -1,5 +1,3 @@
-import { Ref, ref } from 'vue';
-
 interface ScreenShareOptions {
     video: boolean;
     audio: boolean;
@@ -16,12 +14,13 @@ export class ScreenShareService {
         this.context = this.canvas.getContext('2d');
     }
 
-    async startCapture(options: ScreenShareOptions): Promise<void> {
+    async startCapture(options: ScreenShareOptions = { video: true, audio: false }): Promise<MediaStream> {
         try {
             this.stream = await navigator.mediaDevices.getDisplayMedia({
                 video: options.video,
                 audio: options.audio,
             });
+            return this.stream;
         } catch (error) {
             console.error('Error starting screen capture:', error);
             throw error;
@@ -39,21 +38,38 @@ export class ScreenShareService {
         return this.stream;
     }
 
-    startSnapshots(intervalMs: number, quality: number, targetResolution: { width: number; height: number }, onFrame: (frame: Blob) => void): void {
+    startSnapshots(intervalMs: number, quality: number, resolution: string | { width: number; height: number }, onFrame: (frame: Blob) => void): void {
         if (!this.stream || !this.context) {
             throw new Error('Stream is not active or context is not available');
+        }
+
+        // Handle resolution parameter
+        let targetResolution = { width: 1920, height: 1080 };
+        if (typeof resolution === 'object') {
+            targetResolution = resolution;
+        } else if (resolution === '720p') {
+            targetResolution = { width: 1280, height: 720 };
+        } else if (resolution === '480p') {
+            targetResolution = { width: 854, height: 480 };
         }
 
         this.canvas.width = targetResolution.width;
         this.canvas.height = targetResolution.height;
 
+        // Get video element from stream
+        const video = document.createElement('video');
+        video.srcObject = this.stream;
+        video.play();
+
         this.intervalId = window.setInterval(() => {
-            this.context.drawImage(this.stream, 0, 0, this.canvas.width, this.canvas.height);
-            this.canvas.toBlob((blob) => {
-                if (blob) {
-                    onFrame(blob);
-                }
-            }, 'image/jpeg', quality);
+            if (this.context && video.videoWidth > 0) {
+                this.context.drawImage(video, 0, 0, this.canvas.width, this.canvas.height);
+                this.canvas.toBlob((blob) => {
+                    if (blob) {
+                        onFrame(blob);
+                    }
+                }, 'image/jpeg', quality);
+            }
         }, intervalMs);
     }
 
@@ -64,3 +80,7 @@ export class ScreenShareService {
         }
     }
 }
+
+// Create a singleton instance
+const screenShareService = new ScreenShareService();
+export default screenShareService;
